Remove old behavior page code
