+++ packages/app/src/domain/behavior/behavior-table-tile.tsx
@@ -1,199 +1,283 @@
-import {
-  NationalBehaviorValue,
-  RegionalBehaviorValue,
-} from '@corona-dashboard/common';
 import css from '@styled-system/css';
-import { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import { isDefined, isPresent } from 'ts-is-present';
+import ChevronIcon from '~/assets/chevron.svg';
 import { Box } from '~/components/base';
 import { PercentageBar } from '~/components/percentage-bar';
 import { Tile } from '~/components/tile';
-import { Heading } from '~/components/typography';
+import { Heading, InlineText, Text } from '~/components/typography';
+import { BehaviorFormatted } from '~/domain/behavior/logic/use-format-and-sort-behavior';
 import { useIntl } from '~/intl';
-import {
-  BehaviorIdentifier,
-  behaviorIdentifiers,
-  BehaviorTrendType,
-  BehaviorType,
-} from './behavior-types';
+import { colors } from '~/style/theme';
+import { asResponsiveArray } from '~/style/utils';
 import { BehaviorIcon } from './components/behavior-icon';
 import { BehaviorTrend } from './components/behavior-trend';
-import { BehaviorTypeControl } from './components/behavior-type-control';
+import scrollIntoView from 'scroll-into-view-if-needed';
+import { BehaviorIdentifier } from './logic/behavior-types';
 
-type BehaviorValue = NationalBehaviorValue | RegionalBehaviorValue;
-
-interface BehaviorTileProps {
-  behavior: BehaviorValue;
+interface BehaviorTableTileProps {
   title: string;
-  introduction: Record<BehaviorType, string>;
-  footer: Record<BehaviorType, string>;
-  footerAsterisk: Record<BehaviorType, string>;
-}
-
-export interface BehaviorFormatted {
-  id: BehaviorIdentifier;
   description: string;
-  percentage: number;
-  trend?: BehaviorTrendType;
-}
-
-const HeaderCell = styled.th(
-  css({
-    px: 3,
-    py: 2,
-    textAlign: 'left',
-  })
-);
-
-const Cell = styled.td((x) =>
-  css({
-    color: x.color,
-    borderBottom: '1px solid',
-    borderBottomColor: 'lightGray',
-    px: 3,
-    py: 2,
-  })
-);
-
-/* Format raw list of behaviors into list for compliance or for support */
-function useFormatBehaviorType(
-  behavior: BehaviorValue,
-  type: BehaviorType
-): BehaviorFormatted[] {
-  const { siteText } = useIntl();
-
-  return behaviorIdentifiers
-    .map((identifier) => {
-      const percentage = behavior[
-        `${identifier}_${type}` as keyof BehaviorValue
-      ] as number | null;
-      const trend = (behavior[
-        `${identifier}_${type}_trend` as keyof BehaviorValue
-      ] ?? undefined) as BehaviorTrendType | null;
-
-      return isPresent(percentage)
-        ? {
-            id: identifier,
-            description: siteText.gedrag_onderwerpen[identifier],
-            percentage,
-            trend: trend || undefined,
-          }
-        : undefined;
-    })
-    .filter(isDefined);
-}
-
-/* Sort lists of formatted compliance and support behaviors */
-function sortBehavior(
-  compliance: BehaviorFormatted[],
-  support: BehaviorFormatted[]
-): {
+  complianceExplanation: string;
+  supportExplanation: string;
   sortedCompliance: BehaviorFormatted[];
   sortedSupport: BehaviorFormatted[];
-} {
-  /* Sort compliance on percentage, decreasing; percentage can be null */
-  const sortedCompliance = compliance.sort(
-    (a, b) => (b.percentage ?? 0) - (a.percentage ?? 0)
-  );
-
-  /* Sort support based on the order of compliance */
-  const sortedSupport = support.sort((a, b) => {
-    return (
-      sortedCompliance.findIndex((x) => x.id === a.id) -
-      sortedCompliance.findIndex((x) => x.id === b.id)
-    );
-  });
-
-  return { sortedCompliance, sortedSupport };
-}
-
-function useFormatAndSortBehavior(behavior: BehaviorValue): {
-  sortedCompliance: BehaviorFormatted[];
-  sortedSupport: BehaviorFormatted[];
-} {
-  const compliance = useFormatBehaviorType(behavior, 'compliance');
-  const support = useFormatBehaviorType(behavior, 'support');
-
-  return sortBehavior(compliance, support);
+  annotation: string;
+  setCurrentId: React.Dispatch<React.SetStateAction<BehaviorIdentifier>>;
+  scrollRef: { current: HTMLDivElement | null };
 }
 
 export function BehaviorTableTile({
-  behavior,
   title,
-  introduction,
-  footer,
-  footerAsterisk,
-}: BehaviorTileProps) {
-  const { sortedCompliance, sortedSupport } =
-    useFormatAndSortBehavior(behavior);
-  const [behaviorType, setBehaviorType] = useState<BehaviorType>('compliance');
-
-  const { siteText, formatPercentage } = useIntl();
+  description,
+  complianceExplanation,
+  supportExplanation,
+  sortedCompliance,
+  sortedSupport,
+  annotation,
+  setCurrentId,
+  scrollRef,
+}: BehaviorTableTileProps) {
+  const { siteText } = useIntl();
   const commonText = siteText.gedrag_common;
 
   return (
     <Tile>
       <Heading level={3}>{title}</Heading>
-      <Box display="flex" justifyContent="start">
-        <BehaviorTypeControl value={behaviorType} onChange={setBehaviorType} />
+      <Box maxWidth="maxWidthText">
+        <Text mb={4}>{description}</Text>
       </Box>
-
-      <p>{introduction[behaviorType]}</p>
-
-      <div css={css({ overflow: 'auto' })}>
-        <table css={css({ width: '100%', borderCollapse: 'collapse' })}>
+      <Box display="flex" flexWrap="wrap" mb={{ _: 2, xs: 4 }}>
+        <Box mr={3} mb={1}>
+          <ExplanationBox background={colors.data.cyan} />
+          {complianceExplanation}
+        </Box>
+        <Box>
+          <ExplanationBox background={colors.data.yellow} />
+          {supportExplanation}
+        </Box>
+      </Box>
+      <Box overflow="auto" mb={3}>
+        <StyledTable>
           <thead>
             <tr>
-              <HeaderCell colSpan={2}>
+              <HeaderCell
+                css={css({
+                  width: asResponsiveArray({
+                    _: 200,
+                    sm: 300,
+                    xl: 400,
+                  }),
+                })}
+              >
+                {commonText.basisregels.header_basisregel}
+              </HeaderCell>
+              <HeaderCell
+                css={css({
+                  width: asResponsiveArray({
+                    _: 100,
+                    sm: 100,
+                    xl: 150,
+                  }),
+                })}
+              >
                 {commonText.basisregels.header_percentage}
               </HeaderCell>
-              <th />
-              <HeaderCell>
-                <span
-                  css={css({
-                    display: 'inline-block',
-                    minWidth: [180, 200, 250, 180],
-                  })}
-                >
-                  {commonText.basisregels.header_basisregel}
-                </span>
-              </HeaderCell>
-              <HeaderCell>
-                <span css={css({ display: 'inline-block', minWidth: 100 })}>
-                  {commonText.basisregels.header_trend}
-                </span>
+              <HeaderCell
+                css={css({
+                  width: 125,
+                })}
+              >
+                {commonText.basisregels.header_trend}
               </HeaderCell>
             </tr>
           </thead>
           <tbody>
-            {(behaviorType === 'compliance'
-              ? sortedCompliance
-              : sortedSupport
-            ).map((behavior) => (
+            {sortedCompliance.map((behavior, index) => (
               <tr key={behavior.id}>
-                <Cell>{formatPercentage(behavior.percentage)}%</Cell>
-                <Cell color="data.primary">
-                  <PercentageBar percentage={behavior.percentage} />
-                </Cell>
-                <Cell>
-                  <Box minWidth={32}>
-                    <BehaviorIcon name={behavior.id} />
+                <Cell
+                  css={css({
+                    minWidth: asResponsiveArray({ _: '60vw', sm: 300 }),
+                  })}
+                >
+                  <Box display="flex" mr={2}>
+                    <Box minWidth={32} color="black" pr={2} display="flex">
+                      <BehaviorIcon name={behavior.id} size={25} />
+                    </Box>
+                    <DescriptionWithIcon
+                      description={behavior.description}
+                      id={behavior.id}
+                      setCurrentId={setCurrentId}
+                      scrollRef={scrollRef}
+                    />
                   </Box>
                 </Cell>
-                <Cell>{behavior.description}</Cell>
-                <Cell>
-                  <BehaviorTrend trend={behavior.trend} />
+                <Cell css={css({ minWidth: 200 })}>
+                  <PercentageBarWithNumber
+                    percentage={behavior.percentage}
+                    color={colors.data.cyan}
+                  />
+                  <PercentageBarWithNumber
+                    percentage={sortedSupport[index].percentage}
+                    color={colors.data.yellow}
+                  />
+                </Cell>
+                <Cell css={css({ minWidth: 125 })}>
+                  <Box display="flex" flexDirection="column">
+                    <BehaviorTrend trend={behavior.trend} color={colors.body} />
+                    <BehaviorTrend
+                      trend={sortedSupport[index].trend}
+                      color={colors.body}
+                    />
+                  </Box>
                 </Cell>
               </tr>
             ))}
           </tbody>
-        </table>
-      </div>
-      <p css={css({ color: 'annotation' })}>{footer[behaviorType]}</p>
-      <p css={css({ color: 'annotation', m: 0 })}>
-        {footerAsterisk[behaviorType]}
-      </p>
+        </StyledTable>
+      </Box>
+      <Box maxWidth="maxWidthText">
+        <Text color="annotation">{annotation}</Text>
+      </Box>
     </Tile>
   );
 }
+
+/**
+ * Render every word in a span and add the chevron to the last word.
+ * this is for the word wrapping when the screen gets smaller.
+ */
+function DescriptionWithIcon({
+  description,
+  id,
+  setCurrentId,
+  scrollRef,
+}: {
+  description: string;
+  id: BehaviorIdentifier;
+  setCurrentId: React.Dispatch<React.SetStateAction<BehaviorIdentifier>>;
+  scrollRef: { current: HTMLDivElement | null };
+}) {
+  const splittedWords = description.split(' ');
+
+  const buttonClickHandler = () => {
+    scrollIntoView(scrollRef.current as Element);
+    setCurrentId(id);
+  };
+
+  return (
+    <Button onClick={buttonClickHandler}>
+      {splittedWords.map((word, index) => (
+        <InlineText
+          key={index}
+          css={css({
+            whiteSpace: 'pre-wrap',
+            fontFamily: 'body',
+            fontSize: '1rem',
+          })}
+        >
+          {splittedWords.length - 1 === index ? (
+            <InlineText css={css({ display: 'flex', position: 'relative' })}>
+              {word}
+              <Box position="absolute" right={-14} top={0}>
+                <ChevronIcon width="7px" />
+              </Box>
+            </InlineText>
+          ) : (
+            `${word} `
+          )}
+        </InlineText>
+      ))}
+    </Button>
+  );
+}
+
+function PercentageBarWithNumber({
+  percentage,
+  color,
+}: {
+  percentage: number;
+  color: string;
+}) {
+  const { formatPercentage } = useIntl();
+  return (
+    <Box
+      color={color}
+      display="flex"
+      alignItems="center"
+      pr={{ _: 2, sm: 2, lg: 4, xl: 5 }}
+    >
+      <InlineText fontWeight="bold" color="black" pr={2}>
+        {`${formatPercentage(percentage)}%`}
+      </InlineText>
+      <PercentageBar percentage={percentage} height="8px" />
+    </Box>
+  );
+}
+
+const ExplanationBox = styled.div<{ background: string }>((x) =>
+  css({
+    height: '17px',
+    width: '17px',
+    background: x.background,
+    float: 'left',
+    mt: '3px',
+    mr: 1,
+    borderRadius: '3px',
+  })
+);
+
+const StyledTable = styled.table(
+  css({
+    borderCollapse: 'collapse',
+    width: '100%',
+  })
+);
+
+const HeaderCell = styled.th(
+  css({
+    textAlign: 'left',
+  })
+);
+
+const Cell = styled.td(
+  css({
+    borderBottom: '1px solid',
+    borderBottomColor: 'lightGray',
+    p: 0,
+    py: 2,
+  })
+);
+
+const Button = styled.button(
+  css({
+    appearance: 'none',
+    background: 'unset',
+    border: 0,
+    display: 'flex',
+    alignItems: 'center',
+    textAlign: 'left',
+    cursor: 'pointer',
+    p: 0,
+    m: 0,
+    pr: 3,
+    flexWrap: 'wrap',
+
+    '&:focus': {
+      borderColor: 'lightGray',
+      outline: '2px dotted',
+      outlineColor: 'blue',
+    },
+
+    '&:hover': {
+      span: {
+        color: 'data.primary',
+        textDecoration: 'underline',
+      },
+
+      svg: {
+        color: 'data.primary',
+      },
+    },
+  })
+);
